project redirects to active module
