+++ adhocracy4/projects/mixins.py
@@ -1,23 +1,17 @@
 from django.views import generic
 
+from adhocracy4.modules.views import ModuleRedirectView
 
-class PhaseDispatchMixin(generic.DetailView):
+
+class ModuleDispatchMixin(generic.DetailView):
     def dispatch(self, request, *args, **kwargs):
         kwargs['project'] = self.get_object()
-        return self._view_by_phase()(request, *args, **kwargs)
-
-    def _view_by_phase(self):
-        """
-        Choose the appropriate view for the current active phase.
-        """
         project = self.get_object()
-
-        if project.active_phase:
-            return project.active_phase.view.as_view()
-        elif project.past_phases:
-            return project.past_phases[0].view.as_view()
+        if project.last_active_module:
+            view = ModuleRedirectView.as_view()
         else:
-            return super().dispatch
+            view = super().dispatch
+        return view(request, *args, **kwargs)
 
 
 class ProjectMixin(generic.base.ContextMixin):
